+++ app/Http/Controllers/Admin/IndexController.php
@@ -9,6 +9,6 @@ class IndexController extends Controller
 {
     public function index()
     {
-        return view('admin.layout.index');
+        return view('admin.index');
     }
 }
