+++ resources/views/admin/index.blade.php
